Made concurrent uniform and preferentail

+++ src/__main__.py
@@ -1,9 +1,10 @@
 import sys
+import os
 import argparse
 from multiprocessing.context import Process
 
 from src.analysis import *
-from src.data import PreferentialContext, UniformContext, Context
+from src.data import PreferentialContext, UniformZipfContext, Context
 from src.simulation import Simulation
 
 
@@ -13,32 +14,69 @@ def parse():
     parser.add_argument('--people', '-p', metavar='people', type=int, help='Quantity of people in the distribution',
                         default=200)
     parser.add_argument('-s', metavar='s', type=float, help='s constant of Zipf Distribution', default=1.4)
+    parser.add_argument('--threads', metavar='threads', type=int, help='Number of threads for concurrency',
+                        default=4)
     parser.add_argument('--max', metavar='max_msgs', type=int, help='Maximum amount of messages from one only author',
                         default=15)
     parser.add_argument('--seed', metavar='seed', type=int, help='Number of seeds', default=50)
-    parser.add_argument('--sweight', metavar='weight', type=int, help='Value of weight for each signature', default=1)
-    parser.add_argument('--fweight', metavar='weight', type=int, help='Value of weight for each signature', default=1)
-    parser.add_argument('--siweight', metavar='weight', type=int, help='Value of weight for each signature', default=1)
-    parser.add_argument('--fiweight', metavar='weight', type=int, help='Value of weight for each signature', default=200)
     parser.add_argument('--output', '-o', metavar='output', type=str, help='output dir', default='./output/best/')
-    return parser.parse_args()
+
+    subparsers = parser.add_subparsers()
+    subparsers.required = True
+    subparsers.dest = 'command'
+
+    pref = subparsers.add_parser('pref')
+    pref.set_defaults(func=main_preferential)
+    pref.add_argument('--sweight', metavar='weight', type=int, help='Value of weight for each signature', default=1)
+    pref.add_argument('--fweight', metavar='weight', type=int, help='Value of weight for each signature', default=1)
+    pref.add_argument('--siweight', metavar='weight', type=int, help='Value of weight for each signature', default=1)
+    pref.add_argument('--fiweight', metavar='weight', type=int, help='Value of weight for each signature', default=200)
+
+    unif = subparsers.add_parser('unif')
+    unif.set_defaults(func=main_uniform)
+    args = parser.parse_args()
+    args.func(args)
+
+def main_uniform(args):
+    PEOPLE = args.people
+    K = args.k
+    SEED_RANGE = args.seed
+    context: UniformZipfContext= UniformZipfContext(people=PEOPLE, k=K, max_msg=args.max, s=args.s)
+    NUM_THREADS = args.threads
+    output = args.output + f'{SEED_RANGE}-{K}'
+    filename: str = f'{output}/iw1-w0.csv'
+    writer = open(filename, mode='w+')
+    print('seed', 'initial-weight', 'weight', 'none-anonymity',
+            'mean', 'median', 'desviation', 'min', 'max', 'range',
+            'person1', 'person5', 'person10', 'person15',
+            sep=',', file=writer)
+    writer.close()
+    process = {}
+    for m in range(NUM_THREADS):
+        start = int(SEED_RANGE/NUM_THREADS * m)
+        finish = int(SEED_RANGE/NUM_THREADS * (m + 1))
+        process[m] = Process(target=thread_calculation, args=(start, finish, 1, context, 0, filename,))
+        process[m].start()
+    for m in range(NUM_THREADS):
+        process[m].join()
 
 
-def main():
-    args = parse()
+def main_preferential(args):
     PEOPLE = args.people
     K = args.k
     SEED_RANGE = args.seed
     SW, FW = args.sweight, args.fweight
     SIW, FIW = args.siweight, args.fiweight
     output = args.output + f'{SEED_RANGE}-{K}'
-    context: PreferentialContext = PreferentialContext(people=PEOPLE, k=K, max_msg=args.max, s=args.s, initial_weight=1,
-                                           weight=1)
-    NUM_THREADS = 4
+    context: PreferentialContext = PreferentialContext(people=PEOPLE, k=K, max_msg=args.max, s=args.s, initial_weight=args.sweight,
+                                           weight=args.fweight)
+    NUM_THREADS = args.threads
     for w in range(SW, FW + 1):
         for iw in range(SIW, FIW+1):
             print(iw, w)
             filename: str = f'{output}/iw{iw}-w{w}.csv'
+            if os.path.isfile(filename):
+                continue
             writer = open(filename, mode='w+')
             print('seed', 'initial-weight', 'weight', 'none-anonymity',
                   'mean', 'median', 'desviation', 'min', 'max', 'range',
@@ -79,4 +117,4 @@ def thread_calculation(start, finish, iw, context, w, filename):
 
 
 if __name__ == '__main__':
-    main()
+    parse()
